+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod core;
+pub mod db;
 pub mod domain;
 pub mod services;
 pub mod utils;
