+++ src/services/mod.rs
@@ -1,2 +1,1 @@
-pub mod db;
 pub mod registry;
